fetching current cart and putting it on state seems to work

+++ server/db/models/product.js
@@ -25,7 +25,7 @@ const Product = db.define('product', {
     defaultValue: 1
   },
   price: {
-    type: Sequelize.DECIMAL,
+    type: Sequelize.SMALLINT,
     validate: {
       min: 0
     },
